moving some custom logic to new groupby.py file
