+++ python/dask_cudf/dask_cudf/core.py
@@ -1,4 +1,5 @@
 # Copyright (c) 2018-2020, NVIDIA CORPORATION.
+import math
 import warnings
 from distutils.version import LooseVersion
 from operator import getitem
@@ -15,7 +16,6 @@ from dask.context import _globals
 from dask.core import flatten
 from dask.dataframe.core import (
     Scalar,
-    _concat,
     handle_out,
     map_partitions,
     new_dd_object,
@@ -28,7 +28,7 @@ from dask.utils import M, OperatorMethodMixin, derived_from, funcname
 import cudf
 from cudf import _lib as libcudf
 
-from dask_cudf import sorting
+from dask_cudf import groupby, sorting
 
 DASK_VERSION = LooseVersion(dask.__version__)
 
@@ -355,11 +355,19 @@ class DataFrame(_Frame, dd.core.DataFrame):
         gb_cols: list,
         agg_list: list,
         split_every=8,
-        split_out=8,
+        split_out=None,
         dropna=True,
         out_to_host=False,
+        sep="___",
     ):
+        # Deal with default split_out and split_every params
+        if split_every is False:
+            split_every = self.npartitions
+        split_every = split_every or 8
+        split_out = split_out or 1
 
+        # Only support basic aggs and mean
+        # TODO: Support "std" and "var"
         _supported = {"count", "mean", "sum", "min", "max"}
         if not set(agg_list).issubset(_supported):
             raise ValueError(
@@ -367,17 +375,15 @@ class DataFrame(_Frame, dd.core.DataFrame):
             )
 
         dsk = {}
-        sep = "____"
         token = tokenize(self, gb_cols, agg_list)
-        level_1_name = "level_1-" + token
-        split_name = "groupby_split-" + token
-        level_2_name = "level_2-" + token
+        partition_agg_name = "groupby_partition_agg-" + token
+        tree_reduce_name = "groupby_tree_reduce-" + token
         gb_agg_name = "groupby_agg-" + token
         for p in range(self.npartitions):
             # Perform groupby aggregation on each partition.
             # Split each result into `split_out` chunks (by hashing `gb_cols`)
-            dsk[(level_1_name, p)] = (
-                _top_level_groupby,
+            dsk[(partition_agg_name, p)] = (
+                groupby._groupby_partition_agg,
                 (self._name, p),
                 gb_cols,
                 agg_list,
@@ -388,30 +394,47 @@ class DataFrame(_Frame, dd.core.DataFrame):
             )
             # Pick out each chunk using `getitem`
             for s in range(split_out):
-                dsk[(split_name, p, s)] = (getitem, (level_1_name, p), s)
+                dsk[(tree_reduce_name, p, s, 0)] = (
+                    getitem,
+                    (partition_agg_name, p),
+                    s,
+                )
 
-        # For each split, aggregate result for all partitions.
-        # TODO: Use tree reduction here
-
-        # tree_height =
-
+        # Build reduction tree
+        parts = self.npartitions
+        widths = [parts]
+        while parts > 1:
+            parts = math.ceil(parts / split_every)
+            widths.append(parts)
+        height = len(widths)
         for s in range(split_out):
-            dsk[(level_2_name, s)] = (
-                _mid_level_groupby,
-                [(split_name, p, s) for p in range(self.npartitions)],
-                gb_cols,
-                agg_list,
-                split_out,
-                dropna,
-                out_to_host,
-                sep,
-            )
+            for depth in range(1, height):
+                for group in range(widths[depth]):
+
+                    p_max = widths[depth - 1]
+                    lstart = split_every * group
+                    lstop = min(lstart + split_every, p_max)
+                    node_list = [
+                        (tree_reduce_name, p, s, depth - 1)
+                        for p in range(lstart, lstop)
+                    ]
+
+                    dsk[(tree_reduce_name, group, s, depth)] = (
+                        groupby._tree_node_agg,
+                        node_list,
+                        gb_cols,
+                        agg_list,
+                        split_out,
+                        dropna,
+                        out_to_host,
+                        sep,
+                    )
 
         # Final output partitions
         for s in range(split_out):
             dsk[(gb_agg_name, s)] = (
-                _finalize_gb_agg,
-                (level_2_name, s),
+                groupby._finalize_gb_agg,
+                (tree_reduce_name, 0, s, height - 1),
                 gb_cols,
                 agg_list,
                 sep,
@@ -423,98 +446,6 @@ class DataFrame(_Frame, dd.core.DataFrame):
             gb_agg_name, dsk, dependencies=[self]
         )
         return new_dd_object(graph, gb_agg_name, _meta, divisions)
-
-
-def _make_name(*args, sep="_"):
-    _args = (arg for arg in args if arg != "")
-    return sep.join(_args)
-
-
-def _top_level_groupby(
-    df, gb_cols, agg_list, split_out, dropna, out_to_host, sep
-):
-    _agg_list = set()
-    for agg in agg_list:
-        if agg == "mean":
-            _agg_list.add("count")
-            _agg_list.add("sum")
-        else:
-            _agg_list.add(agg)
-    _agg_list = list(_agg_list)
-
-    gb = df.groupby(gb_cols, dropna=dropna, as_index=False).agg(_agg_list)
-    gb.columns = [_make_name(*name, sep=sep) for name in gb.columns]
-    output = {}
-    for j, split in enumerate(
-        gb.partition_by_hash(gb_cols, split_out, keep_index=False)
-    ):
-        if out_to_host:
-            output[j] = split.to_pandas()
-        else:
-            output[j] = split
-    del gb
-    return output
-
-
-def _mid_level_groupby(
-    dfs, gb_cols, agg_list, split_out, dropna, out_to_host, sep
-):
-    df = _concat(dfs, ignore_index=True)
-    if out_to_host:
-        df.reset_index(drop=True, inplace=True)
-        df = cudf.from_pandas(df)
-    agg_dict = {}
-    for col in df.columns:
-        if col in gb_cols:
-            continue
-        agg = col.split(sep)[-1]
-        if agg in ("count", "sum"):
-            agg_dict[col] = ["sum"]
-        elif agg in ("min", "max"):
-            agg_dict[col] = [agg]
-        else:
-            raise ValueError(f"Unexpected aggregation: {agg}")
-
-    gb = df.groupby(gb_cols, dropna=dropna, as_index=False).agg(agg_dict)
-
-    # Don't include the last aggregation in the column names
-    gb.columns = [_make_name(*name[:-1], sep=sep) for name in gb.columns]
-    return gb
-
-
-def _finalize_gb_agg(gb, gb_cols, agg_list, sep):
-
-    # Deal with "mean"
-    if "mean" in agg_list:
-        for col in gb.columns:
-            if col in gb_cols:
-                continue
-            name = col.split(sep)
-            # import pdb; pdb.set_trace()
-            if name[-1] == "sum":
-                mean_name = _make_name(*(name[:-1] + ["mean"]), sep=sep)
-                count_name = _make_name(*(name[:-1] + ["count"]), sep=sep)
-                sum_name = _make_name(*name, sep=sep)
-                gb[mean_name] = gb[sum_name] / gb[count_name]
-                if "sum" not in agg_list:
-                    gb.drop(columns=[sum_name], inplace=True)
-                if "count" not in agg_list:
-                    gb.drop(columns=[count_name], inplace=True)
-
-    # Unflatten column names
-    col_array = []
-    agg_array = []
-    for col in gb.columns:
-        if col in gb_cols:
-            col_array.append(col)
-            agg_array.append("")
-        else:
-            name, agg = col.split(sep)
-            col_array.append(name)
-            agg_array.append(agg)
-    gb.columns = pd.MultiIndex.from_arrays([col_array, agg_array])
-
-    return gb
 
 
 def sum_of_squares(x):
